Add script to fetch and validate latest images
The script gets the list of releases from the CopperheadOS site and
pulls down the latest image for a given device or devices.  After bing
fetched the signature is downloaded and validated.

GPG keys are not fetched automatically, so if the user doesn't have the
public key of the signer they will see a "valid signature" message, but
no indication of trust.

The implementation to work out the image file name is based on the
javascript used by the CopperheadOS site.